python/bigquery/requirements.txt
google-cloud-bigquery_datatransfer==3.15.2